+++ CHANGELOG.md
@@ -16,6 +16,8 @@ Cinecred Changelog
 
 ### UI Fixes
 
+- While the crash dialog is open, launching another instance no longer brings up
+  a new welcome window.
 - Fixed rare crashes caused by numerical inaccuracies when shifting a low-res
   page preview image.
 - Clicking a link in the missing Ghostscript dialog no longer causes a crash.